requirements.txt
pytest>=7.0.0
ragas>=0.1.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langsmith>=0.1.0
requests>=2.31.0 